fastapi/requirements.txt
# --- Framework principal ---
fastapi==0.119.0
uvicorn[standard]==0.37.0

# --- Supabase ---
supabase==2.22.0

# --- Base de datos y ORM ---
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
alembic==1.13.2

# --- Variables de entorno ---
python-dotenv==1.0.1

# --- Autenticación y seguridad ---
python-jose==3.3.0       # para JWT
passlib[bcrypt]==1.7.4   # para hashear contraseñas/códigos
email-validator==2.1.0   # para validar correos
itsdangerous==2.1.2      # para generar tokens temporales (para códigos por correo)

# --- Envío de correos ---
aiosmtplib==2.0.2        # envío asíncrono
jinja2==3.1.4             # plantillas HTML para correos

# --- Validación y esquemas ---
pydantic==2.12.2
pydantic-settings==2.5.2

# --- Utilidades ---
requests==2.32.3
python-multipart==0.0.9   # para subir imágenes (fotos de especies)
Pillow==10.3.0            # procesamiento básico de imágenes

# --- Testing ---
pytest==8.3.3
httpx==0.27.0             # pruebas de endpoints

# --- Despliegue / Servidor (opcional) ---
gunicorn==23.0.0          # para servidor de producción (opcional)